Version 0.1.0
Creating the initial files of the project, setting up the database with the 'users' table, and developing the functionalities and business logic related to the user

Creating the base files for the project and configuration files, such as the 'knexfile.js' file to set up the 'knex' library and the 'jest.config.js' file to configure 'Jest' for application testing. Creating the database using 'knex' and generating the user table using migrations. Developing the business logic and functionalities related to application users, utilizing the concept of controllers and middlewares. Finally, establishing API routes corresponding to user functionalities and implementing the logic for initiating a session in the application

Creating the tables for dishes and ingredients and developing the business logic and functionalities related to dishes and their ingredients in the application.

Creating two more tables in the database to represent the models present in the application, establishing the 'dishes' table to store registered dishes. This table includes the fields 'id', 'dish_image', 'name', 'description', 'price', and 'category', Additionally, establishing the 'dishes_ingredients' table to store dish ingredients, with fields including 'id', 'dish_id', and 'name.' Developing functionalities and building business logic related to dishes and their ingredients.Creating the 'DishesController' and 'IngredientsController' classes to manage and provide functionalities for the application routes. Implementing services for creating, updating, displaying, listing, and deleting dishes within the 'Dishes' folder under 'services.' Also, establishing services for creating, updating, and listing ingredients within the 'Ingredients' folder under 'services.' Creating routes for dishes and ingredients through the 'dishes.routes.js' and 'ingredients.routes.js' files. Finally, implementing logic to save dish images in the application and display them. This involves creating the 'tmp' folder to temporarily store images and the 'uploads' folder within 'tmp' for permanent storage.

Updating the 'auth.config.js' and 'SessionsController.js' files and creating a route to validate the user

Updating the 'auth.config.js' file by changing the expiration time of the expiration token to one day, and modifying the 'SessionsController' class to do the same with the authentication cookie. Finally, creating a route to validate if the logged-in user really exists in the application

Updating the password and price regex

Creating the cookie management route and preparing the application for deployment

Creating the final files of the first version of the API, creating a new route to manage the cookies sent to the Front-End, and a new controller containing the logic to remove a cookie. Finally, preparing the application for deployment, generating the 'ecosystem.config.cjs' file to configure the 'pm2' library, which will execute the application in the development environment, ensuring that the application is always available.

Updating Cookie's destination

Creating the README.md file and finishing the first version of application