Creating the initial files of the project, setting up the database with the 'users' table, and developing the functionalities and business logic related to the user
Creating the base files for the project and configuration files, such as the 'knexfile.js' file to set up the 'knex' library and the 'jest.config.js' file to configure 'Jest' for application testing. Creating the database using 'knex' and generating the user table using migrations. Developing the business logic and functionalities related to application users, utilizing the concept of controllers and middlewares. Finally, establishing API routes corresponding to user functionalities and implementing the logic for initiating a session in the application