Creating the cookie management route and preparing the application for deployment
Creating the final files of the first version of the API, creating a new route to manage the cookies sent to the Front-End, and a new controller containing the logic to remove a cookie. Finally, preparing the application for deployment, generating the 'ecosystem.config.cjs' file to configure the 'pm2' library, which will execute the application in the development environment, ensuring that the application is always available.